+++ onward/app/controllers/SeriesController.scala
@@ -17,28 +17,10 @@ import play.api.libs.json.{JsArray, Json}
 import play.api.mvc._
 import services.CollectionConfigWithId
 import views.support.FaciaToMicroFormat2Helpers._
-import models.{OnwardCollection, OnwardItem}
+import models.{Series, SeriesStoriesDCR}
 
 import scala.concurrent.Future
 import scala.concurrent.duration._
-
-case class Series(id: String, tag: Tag, trails: RelatedContent) {
-  lazy val displayName = tag.id match {
-    case "commentisfree/commentisfree" => "opinion"
-    case _ => tag.metadata.webTitle
- }
-}
-
-case class SeriesStoriesDCR(id: String, displayname: String, tag: Tag, trails: Seq[OnwardItem])
-
-object SeriesStoriesDCR {
-  implicit val onwardItemWrites = Json.writes[OnwardItem]
-  implicit val seriesStoriesDCRWrites = Json.writes[SeriesStoriesDCR]
-  def fromSeries(series: Series)(implicit request: RequestHeader): SeriesStoriesDCR = {
-    val trails = OnwardCollection.trailsToItems(series.trails.faciaItems)
-    SeriesStoriesDCR(series.id, series.displayName, series.tag, trails)
-  }
-}
 
 class SeriesController(
   contentApiClient: ContentApiClient,
